qemu: Fix cleanups in case vm or remote was not initialized
Current implementation of cleanup doesn't behave well when VM or remote
are not initialized. Additionally add TODO: message to wait until VM
goes down as currently we just execute shutdown and immediately
destroy it using monitor. That's not really a nice behavior.

+++ qemu/run_iozone.py
@@ -26,5 +26,9 @@ class RunIOZoneTest(test.VirtTest):
         self.whiteboard = self.vm.remote.run('iozone -a').stdout
 
     def cleanup(self):
-        self.vm.remote.run('shutdown -h now')
-        self.vm.power_off()
+        if self.vm:
+            if self.vm.remote:
+                self.vm.remote.run('shutdown -h now')
+                # TODO: Wait for machine to go down
+            self.vm.power_off()
+
