+++ qemu/usb_boot.py
@@ -81,5 +81,9 @@ class USBBootTest(test.VirtTest):
             raise exceptions.TestFail('USB boot test found problems in: %s' % " ".join(e_msg))
 
     def cleanup(self):
-        self.vm.remote.run('shutdown -h now')
-        self.vm.power_off()
+        if self.vm:
+            if self.vm.remote:
+                self.vm.remote.run('shutdown -h now')
+                # TODO: Wait for machine to go down
+            self.vm.power_off()
+
